requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
httpx